requirements_prod.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2-binary
drf-spectacular>=0.15.1,<0.16
django-cors-headers>=3.13,<3.14
django-filter==22.1
python-engineio==4.3.0
python-socketio==5.5.0
sanic==21.12.2
Sanic-Cors==2.0.0
sanic-routing==0.7.2
sanic-jwt==1.7.0
sanic-plugin-toolkit==1.2.0
socketIO-client==0.7.2
gunicorn==20.1.0
uvicorn==0.21.1
safetensors==0.3.1
ninja==1.11.1
sentencepiece
Xformers
einops
chromadb
InstructorEmbedding
sentence_transformers
langchain
django-oauth-toolkit
gpustat